adding stuff

+++ Jenkinsfile
@@ -2,17 +2,32 @@ pipeline {
 
     agent any
 
+    enironment {
+        // Declare variable to set environment
+        WorkingEnv = "undetermined"
+    }
+
     stages {
 
          stage('Init') {
 
             steps {
+                echo "You have selected $(GIT_BRANCH) GIT Branch"
+                if (env.GIT_BRANCH =='origin/main') {
 
                 sh '''
                 ssh -i ~/.ssh/id_rsa jenkins@10.154.0.25  << EOF
                 docker rm -f $(docker ps -qa) || true
                 docker network create new-network || true
                 '''
+                }
+                else if (env.GIT_BRANCH == 'origin/test') {
+                    sh '''
+                ssh -i ~/.ssh/id_rsa jenkins@TestIP  << EOF
+                docker rm -f $(docker ps -qa) || true
+                docker network create new-network || true
+                '''
+                }
 
             }
